move txpool statics of size and cycles to poolmap
